+++ corejava/src/org/isis/promise/PromiseException.java
@@ -24,7 +24,7 @@ final class PromiseException<Type> implements Promise<Type> {
 	}
 
 	@Override
-	public void requestArgument(short index, Future<?> parent) {
+	public void requestArgument(int index, Future<?> parent) {
 		assert (false);
 	}
 }
